# src/main/resources/static/openapi-3.0.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GeoCube Web API Documentation",
    "description": "This is a GeoCube API documentation",
    "contact": {
      "name": "Kaixuan Wang(WHU)",
      "email": "wangkaixuan@whu.edu.cn"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://oge.whu.edu.cn/"
    }
  ],
...
  "paths": {
    "/geocube/gdc_api_t19": {
      "get": {
        "tags": [
          "GDC API - Data Access: STAC API/OGC API - Coverages"
        ],
        "summary": "Backend information",
        "description": "Backend information",
        "operationId": "getBackendInfoUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BackendInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        }
      }
    },
    "/geocube/gdc_api_t19/": {
      "get": {
        "tags": [
          "GDC API - Data Access: STAC API/OGC API - Coverages"
        ],
        "summary": "Backend information",
        "description": "Backend information",
        "operationId": "getBackendInfoUsingGET_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BackendInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        }
      }
    },
...
    "/geocube/gdc_api_t19/collections": {
      "get": {
        "tags": [
          "GDC API - Data Access: STAC API/OGC API - Coverages"
        ],
        "summary": "Collections provided",
        "description": "Collections list",
        "operationId": "getCollectionsUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Collections"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        }
      }
    },
    "/geocube/gdc_api_t19/collections/{collectionId}": {
      "get": {
        "tags": [
          "GDC API - Data Access: STAC API/OGC API - Coverages"
        ],
        "summary": "Collection provided",
        "description": "Get the description of a certain collection. Collection is used to describe the descriptive information of a Cube and supports the STAC Cube Extension.",
        "operationId": "getCollectionUsingGET",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "description": "collectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        }
      }
    },
...
    "/geocube/gdc_api_t19/jobs": {
      "get": {
        "tags": [
          "GDC API - Data Process: OGC API - Processes - Part1/Part3"
        ],
        "summary": "Get the status of a job",
        "description": "Return the job status",
        "operationId": "getJobListUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        }
      }
    },
    "/geocube/gdc_api_t19/jobs/{jobId}": {
      "get": {
        "tags": [
          "GDC API - Data Process: OGC API - Processes - Part1/Part3"
        ],
        "summary": "Get the status of a job",
        "description": "Return the job status",
        "operationId": "getStatusUsingGET",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "jobId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        }
      }
    },
    "/geocube/gdc_api_t19/jobs/{jobId}/results": {
      "get": {
        "tags": [
          "GDC API - Data Process: OGC API - Processes - Part1/Part3"
        ],
        "summary": "Get the result of a job",
        "description": "Return the job result",
        "operationId": "getResultsUsingGET",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "jobId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        }
      }
    },
    "/geocube/gdc_api_t19/processes": {
      "get": {
        "tags": [
          "GDC API - Data Process: OGC API - Processes - Part1/Part3"
        ],
        "summary": "Processes provided",
        "description": "Processes list",
        "operationId": "getProcessesUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessesDesc"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        }
      }
    },
    "/geocube/gdc_api_t19/processes/{processId}": {
      "get": {
        "tags": [
          "GDC API - Data Process: OGC API - Processes - Part1/Part3"
        ],
        "summary": "Process description",
        "description": "Process description",
        "operationId": "getProcessDescriptionUsingGET",
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "description": "processId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDesc"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        }
      }
    },
    "/geocube/gdc_api_t19/processes/{processId}/execution": {
      "post": {
        "tags": [
          "GDC API - Data Process: OGC API - Processes - Part1/Part3"
        ],
        "summary": "Execute a process",
        "description": "Return the job information. When `response = collection`, it redirects to the `/collections/{collectionId}` endpoint of OGC API-Coverages",
        "operationId": "executeUsingPOST",
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "description": "processId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "response",
            "in": "query",
            "description": "response",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "processRequestBody",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "processRequestBody"
      }
    }
  },
  "components": {
    "schemas": {
      "Asset": {
        "title": "Asset",
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "BackendInfo": {
        "title": "BackendInfo",
        "type": "object",
        "properties": {
          "api_version": {
            "type": "string"
          },
          "backend_version": {
            "type": "string"
          },
          "conformsTo": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string"
          },
          "endpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Endpoint"
            }
          },
          "gdc_version": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          },
          "stac_version": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "Collection": {
        "title": "Collection",
        "type": "object",
        "properties": {
          "assets": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "cube:dimensions": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Dimension"
            }
          },
          "deprecated": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "extent": {
            "$ref": "#/components/schemas/Extent"
          },
          "id": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "license": {
            "type": "string"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Providers"
            }
          },
          "sci:citation": {
            "type": "string"
          },
          "stac_extensions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "stac_version": {
            "type": "string"
          },
          "summaries": {
            "$ref": "#/components/schemas/CollectionSummaries"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "Collections": {
        "title": "Collections",
        "type": "object",
        "properties": {
          "collections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Collection"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          }
        }
      },
...
      "Dimension": {
        "title": "Dimension",
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "bands",
              "other",
              "spatial",
              "temporal"
            ]
          }
        }
      },
...
      "Endpoint": {
        "title": "Endpoint",
        "type": "object",
        "properties": {
          "methods": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "Input": {
        "title": "Input",
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxOccurs": {
            "type": "integer",
            "format": "int32"
          },
          "metadata": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minOccurs": {
            "type": "integer",
            "format": "int32"
          },
          "schema": {
            "$ref": "#/components/schemas/Schema"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "Link": {
        "title": "Link",
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "rel": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "Output": {
        "title": "Output",
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "schema": {
            "$ref": "#/components/schemas/Schema"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "ProcessDesc": {
        "title": "ProcessDesc",
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "inputs": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Input"
            }
          },
          "jobControlOptions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          },
          "outputTransmission": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "outputs": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Output"
            }
          },
          "title": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "ProcessRequestBody": {
        "title": "ProcessRequestBody",
        "type": "object",
        "properties": {
          "inputs": {
            "type": "object",
            "properties": {}
          },
          "outputs": {
            "type": "object",
            "properties": {}
          },
          "process": {
            "type": "string"
          },
          "response": {
            "type": "string"
          },
          "subscriber": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {}
            }
          }
        }
      },
      "ProcessesDesc": {
        "title": "ProcessesDesc",
        "type": "object",
        "properties": {
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          },
          "processes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessDesc"
            }
          }
        }
      },
...
    }
  }
}